requirements.txt
Pillow==10.2.0
tqdm==4.66.2
ultralytics==8.1.34
torch==2.1.0
torchvision==0.16.0
numpy
ensemble-boxes
torchmetrics>=1.3
pyyaml